+++ src/clients.rs
@@ -323,6 +323,7 @@ impl StateClient for StateStore {
         vk: &ViewKey,
     ) -> Result<Vec<EnrichedNote>, Self::Error> {
         let psk = vk.public_spend_key();
+        self.sync().wait()?;
         let state = self.inner.lock().unwrap();
 
         Ok(state